+++ tests/PermissionTest.php
@@ -2,6 +2,8 @@
 
 namespace TPG\Tests;
 
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
 use TPG\Deadbolt\Exceptions\NoSuchPermissionException;
 use TPG\Deadbolt\Facades\Deadbolt;
 
@@ -123,5 +125,32 @@ class PermissionTest extends TestCase
         Deadbolt::user($user)->revokeAll();
 
         $this->assertTrue(Deadbolt::user($user)->hasNone(Deadbolt::permissions()));
+    }
+
+    /**
+     * @test
+     */
+    public function it_allows_a_different_column_name()
+    {
+        $this->app['config']->set('deadbolt.column', 'rights');
+
+        Schema::drop('users');
+        Schema::create('users', function (Blueprint $table) {
+            $table->bigIncrements('id');
+            $table->string('name');
+            $table->string('email')->unique();
+            $table->timestamp('email_verified_at')->nullable();
+            $table->string('password');
+            $table->text('rights')->nullable();
+            $table->rememberToken();
+            $table->timestamps();
+        });
+
+        $user = $this->user();
+
+        Deadbolt::user($user)->give('articles.create', 'articles.edit')->save();
+
+        $this->assertTrue(Deadbolt::user($user)->has('articles.edit'));
+        $this->assertContains('articles.create', Deadbolt::user($user)->permissions());
     }
 }