+++ tests/User.php
@@ -8,4 +8,8 @@ use Illuminate\Notifications\Notifiable;
 class User extends Authenticatable
 {
     use Notifiable;
+
+    protected $casts = [
+        'permissions' => 'json'
+    ];
 }